// store/search/types/enums/getter-type.ts
export enum GetterType {
    GET_TYPE_MODE_ENABLED = "GET_TYPE_MODE_ENABLED",
    GET_QUERY = "GET_QUERY",
    GET_LOADING = "GET_LOADING",
    GET_HAS_RESULT = "GET_HAS_RESULT",

    GET_LIVE_RESULT_PREVIEW = "GET_LIVE_RESULT_PREVIEW",
    GET_LIVE_RESULT = "GET_LIVE_RESULT",
    GET_LIVE_RESULT_ITEM_BY_ID = "GET_LIVE_RESULT_ITEM_BY_ID",
    GET_LIVE_RESULT_COUNT = "GET_LIVE_RESULT_COUNT",
    GET_LIVE_HAS_RESULT = "GET_LIVE_HAS_RESULT",
    GET_LIVE_HAS_MORE_RESULT = "GET_LIVE_HAS_MORE_RESULT",
    GET_LIVE_PAGING = "GET_LIVE_PAGING",
    GET_LIVE_LOADING = "GET_LIVE_LOADING",
    GET_LIVE_SEARCH_DATE = "GET_LIVE_SEARCH_DATE",

    GET_MOVIES_RESULT_PREVIEW = "GET_MOVIES_RESULT_PREVIEW",
    GET_MOVIES_RESULT = "GET_MOVIES_RESULT",
    GET_MOVIES_RESULT_COUNT = "GET_MOVIES_RESULT_COUNT",
    GET_MOVIES_HAS_RESULT = "GET_MOVIES_HAS_RESULT",
    GET_MOVIES_HAS_MORE_RESULT = "GET_MOVIES_HAS_MORE_RESULT",
    GET_MOVIES_PAGING = "GET_MOVIES_PAGING",
    GET_MOVIES_LOADING = "GET_MOVIES_LOADING",

    GET_SHOWS_RESULT_PREVIEW = "GET_SHOWS_RESULT_PREVIEW",
    GET_SHOWS_RESULT = "GET_SHOWS_RESULT",
    GET_SHOWS_RESULT_COUNT = "GET_SHOWS_RESULT_COUNT",
    GET_SHOWS_HAS_RESULT = "GET_SHOWS_HAS_RESULT",
    GET_SHOWS_HAS_MORE_RESULT = "GET_SHOWS_HAS_MORE_RESULT",
    GET_SHOWS_PAGING = "GET_SHOWS_PAGING",
    GET_SHOWS_LOADING = "GET_SHOWS_LOADING",

    GET_CHANNELS_RESULT_PREVIEW = "GET_CHANNELS_RESULT_PREVIEW",
    GET_CHANNELS_RESULT = "GET_CHANNELS_RESULT",
    GET_CHANNELS_RESULT_COUNT = "GET_CHANNELS_RESULT_COUNT",
    GET_CHANNELS_HAS_RESULT = "GET_CHANNELS_HAS_RESULT",
    GET_CHANNELS_HAS_MORE_RESULT = "GET_CHANNELS_HAS_MORE_RESULT",
    GET_CHANNELS_PAGING = "GET_CHANNELS_PAGING",
    GET_CHANNELS_LOADING = "GET_CHANNELS_LOADING",

    GET_CREDITS_RESULT_PREVIEW = "GET_CREDITS_RESULT_PREVIEW",
    GET_CREDITS_RESULT = "GET_CREDITS_RESULT",
    GET_CREDITS_RESULT_COUNT = "GET_CREDITS_RESULT_COUNT",
    GET_CREDITS_HAS_MORE_RESULT = "GET_CREDITS_HAS_MORE_RESULT",
    GET_CREDITS_PAGING = "GET_CREDITS_PAGING",
    GET_CREDITS_LOADING = "GET_CREDITS_LOADING",
}